% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERIT_functions.R
\name{load_datasets}
\alias{load_datasets}
\title{Load dataset function}
\usage{
load_datasets(filename.genes = "data/genes.txt",
  filename.experiment_ids = "data/experiment_ids.txt",
  filename.foldChange_differentialExpression = "data/m.foldChange_differentialExpression.txt",
  filename.pvalue_differentialExpression = "data/m.pvalue_differentialExpression.txt",
  filename.experiment_condition_tissue_annotation = "data/df.experiment_condition_annotation.txt",
  filename.transcriptionfactor_annotation = "data/df.transcriptionFactorAnnotation.txt",
  filename.geneGroups = "data/df.enzymes_w_metabolic_domains.txt")
}
\description{
This function loads a datasets
}
\examples{
load_datasets()
}
